commit-graph, fuzz: add fuzzer for commit-graph
Break the logic to parse the in-core data in load_commit_graph_one()
into a new function parse_commit_graph().  The latter function
operates on arbitrary buffers, which makes it suitable as a fuzzing
target.

Add fuzz-commit-graph.c, which provides a fuzzing entry point
compatible with libFuzzer (and possibly other fuzzing engines).

Signed-off-by: Josh Steadmon <steadmon@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>